generate: use exponential random search for seed
A linear increase in identifier will, over time, become more expensive in database queries and time as the seed space is exhausted.
By using the failed seed as the upper bound, the number of existence hits will grow with the logarithm of the database size.

Note that this is not a practical performance concern: this code runs in a batch job, and actually generating the puzzle is much more expensive. This change is intended to spread knowledge of exponential backoff techniques.